+++ src/main/java/insane96mcp/runeenchanting/data/provider/RELanguageProvider.java
@@ -21,8 +21,8 @@ public class RELanguageProvider extends LanguageProvider {
                 add(rune.getInfoTranslationKey(), rune.getInfo());
         }
 
-        add("death.attack.runeenchanting.blood_fueled", "%1$s's blood-fueled rune drained too much life essence");
-        add("death.attack.runeenchanting.blood_pact", "%1$s's blood pact was fulfilled");
+        add("death.attack.runeenchanting.blood_fueled", "%1$s's was consumed by blood-fueled rune");
+        add("death.attack.runeenchanting.blood_pact", "%1$s's was consumed by the blood pact");
         add("death.attack.runeenchanting.electrocution", "%1$s was electrocuted");
 
         add("item.runeenchanting.rune", "Rune");